backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == ['\"]__main__['\"]:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "@abstractmethod",
]

[tool.ruff]
line-length = 79
fix = true
select = ["E", "F"]
extend-ignore = ["E501"]

[tool.poetry]
name = "analytics-chat-agent"
version = "0.1.0"
description = "Analytics Chat Agent"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "analytics_chat_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
langchain = "^0.0.350"
langchain-openai = "^0.0.2"
openai = "^1.3.7"
typer = "^0.16.0"
sentence-transformers = "^4.1.0"
qdrant-client = "1.12.1"
google-cloud-bigquery = "^3.17.2"
google-generativeai = "^0.8.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
ruff = "^0.1.9"

[tool.poetry.scripts]
analytics-cli = "analytics_chat_agent.cli.main:app"

